Module Eliom

module Bus : sig ... end

Broadcasting facilities between clients and server.

module Client : sig ... end

Call server side services and change the current page.

module Client_base : sig ... end

module Client_core : sig ... end

module Client_main : sig ... end

module Client_value : sig ... end

module Comet : sig ... end

Handle unsolicited server to client communications.

module Comet_base : sig ... end

module Common : sig ... end

module Common_base : sig ... end

module Config : sig ... end

module Content : sig ... end

This module provides the creation of valid XML content, i.e. XML, SVG, and (X)HTML5.

module Content_core : sig ... end

XML building and deconstructing. Cf. Content_core.Xml.

module Content_functor : sig ... end

module Content_sigs : sig ... end

module Cookies_base : sig ... end

module Cscache : sig ... end

module Eliom_form : sig ... end

module Eliom_lazy : sig ... end

module Eliom_react : sig ... end

Client side type declarations for React event propagation. This module must be linked for events to work properly.

module Eliom_uri : sig ... end

Low-level functions for relative or absolute URL calculation.

module Form_sigs : sig ... end

module Lib : sig ... end

Eliom standard library

module Lib_base : sig ... end

module Mod_cookies : sig ... end

module Mod_dom : sig ... end

Cross browser dom manipulation functions

module Mod_parameters : sig ... end

module Parameter : sig ... end

module Parameter_base : sig ... end

module Parameter_sigs : sig ... end

Ad-hoc runtime type representation for service parameters.

module Process : sig ... end

module Registration : sig ... end

Client-side service registration.

module Registration_sigs : sig ... end

module Request : sig ... end

module Request_info : sig ... end

This module contains the functions you need to get (or set) information about current request.

module Route : sig ... end

module Route_base : sig ... end

module Runtime : sig ... end

module Service : sig ... end

Creation and manipulation of Eliom services.

module Service_base : sig ... end

module Service_sigs : sig ... end

module Shared : sig ... end

This module implements shared (i.e., client-server) versions of the React and ReactiveData libraries.

module Shared_content : sig ... end

module Shared_sigs : sig ... end

module Tools : sig ... end

Helpers for (hierarchical) menu generation in HTML5. See the Eliom manual for more information about menu or hierarchical site.

module Types : sig ... end

module Types_base : sig ... end

module Unwrap : sig ... end

module Wrap : sig ... end