Module Content.Html

Building valid (X)HTML5.

See the Eliom manual for more information on dom semantics vs. functional semantics for HTML5 tree manipulated by client/server application.

type +'a elt

type +'a attrib

type uri = Xml.uri

type 'a form_param

module type T = sig ... end

Tyxml HTML signature (with Eliom's links/forms extensions) shared by F and D, suitable as a functor parameter. Avoids module type of Content.Html.F, which captures wrapped module paths and triggers an OCaml strengthening bug (see ia-reports/2026-04-ocaml-wrapped-mtof-bug.md).

module F : sig ... end

Creation of functional HTML5 content (copy-able but not referable).

module D : sig ... end

Creation of HTML5 content with DOM semantics (referable)

module R : sig ... end

Creation of HTML5 content from React signals. HTML5's trees are automatically updated whenever corresponding signals change.

module C : sig ... end

Creation of HTML5 content from client-side values. This module is available on client side only to make possible to use C-nodes in shared sections.

module Id : sig ... end

Node identifiers

module Custom_data : sig ... end

module To_dom : Js_of_ocaml_tyxml.Tyxml_cast_sigs.TO with type 'a elt = 'a elt

Conversion from HTML5 elts to Javascript DOM elements (<: Js_of_ocaml.Dom_html.element). One conversion function per source type (stressed by the of_ prefix).

module Manip : sig ... end

DOM-like manipulation functions.

module Of_dom : Js_of_ocaml_tyxml.Tyxml_cast_sigs.OF with type 'a elt = 'a elt

Conversion functions from DOM nodes (Js_of_ocaml.Dom_html.element Js_of_ocaml.Js.t) to Eliom nodes (Content.Html.elt).