Module Eliom.Parameter_base

val section : Logs.src

type params = (string * Mod_parameters.param) list

type params' = (string * string) list

type +'a param_name = string

Type of names in a form

type no_param_name

empty type used when it is not possible to use the parameter in a form

type ('a, 'b) binsum =

| Inj1 of 'a

| Inj2 of 'b

type 'an listnames = {

it : 'el 'a. ('an -> 'el -> 'a -> 'a) -> 'el list -> 'a -> 'a;

}

type coordinates = {

abscissa : int;

ordinate : int;

}

type 'a setoneradio = [

| `Set of 'a

| `One of 'a

| `Radio of 'a

]

type 'a oneradio = [

| `One of 'a

| `Radio of 'a

]

type 'a setone = [

| `Set of 'a

| `One of 'a

]

type 'a to_and_of = 'a Common.to_and_of = {

of_string : string -> 'a;

to_string : 'a -> string;

}

type _ atom =

| TFloat : float atom

| TInt : int atom

| TInt32 : int32 atom

| TInt64 : int64 atom

| TNativeint : nativeint atom

| TBool : bool atom

| TString : string atom

val string_of_atom : 'a atom -> 'a -> string

val atom_of_string : 'a atom -> string -> 'a

val to_from_of_atom : 'a atom -> 'a to_and_of

type 'a filter = ('a -> unit) option

type raw = unit

type 'a ocaml = string

type suff = [

| `WithoutSuffix

| `WithSuffix

| `Endsuffix

]

type (_, _) params_type_ =

| TProd : (('a, 'an) params_type_ * ('b, 'bn) params_type_) -> ('a * 'b, 'an * 'bn) params_type_

| TOption : (('a, 'an) params_type_ * bool) -> ('a option, 'an) params_type_

| TList : (string * ('a, 'an) params_type_) -> ('a list, 'an listnames) params_type_

| TSet : ('a, [ `One of 'an ] param_name) params_type_ -> ('a list, [ `Set of 'an ] param_name) params_type_

| TSum : (('a, 'an) params_type_ * ('b, 'bn) params_type_) -> (('a, 'b) binsum, 'an * 'bn) params_type_

| TAtom : (string * 'a atom) -> ('a, [ `One of 'a ] param_name) params_type_

| TCoord : string -> (coordinates, [ `One of coordinates ] param_name) params_type_

| TFile : string -> (Lib.file_info, [ `One of Lib.file_info ] param_name) params_type_

| TUserType : (string * 'a Common.To_and_of_shared.t) -> ('a, [ `One of 'a ] param_name) params_type_

| TTypeFilter : (('a, 'an) params_type_ * 'a filter) -> ('a, 'an) params_type_

| TESuffix : string -> (string list, [ `One of string list ] param_name) params_type_

| TESuffixs : string -> (string, [ `One of string ] param_name) params_type_

| TESuffixu : (string * 'a Common.To_and_of_shared.t) -> ('a, [ `One of 'a ] param_name) params_type_

| TSuffix : (bool * ('s, 'sn) params_type_) -> ('s, 'sn) params_type_

| TUnit : (unit, unit) params_type_

| TAny : ((string * string) list, unit) params_type_

| TConst : string -> (unit, [ `One of unit ] param_name) params_type_

| TNLParams : ('a, 'names) non_localized_params_ -> ('a, 'names) params_type_

| TJson : (string * 'a Deriving_Json.t option) -> ('a, [ `One of 'a ocaml ] param_name) params_type_

| TRaw_post_data : (raw, no_param_name) params_type_

and ('a, 'names) non_localized_params_ = {

name : string;

persistent : bool;

get : 'a option Polytables.key;

post : 'a option Polytables.key;

param : ('a, 'names) params_type_;

}

type ('a, +'suff, 'an) non_localized_params = ('a, 'an) non_localized_params_ constraint 'suff = [< suff ]

type ('a, +'suff, 'an) params_type = ('a, 'an) params_type_ constraint 'suff = [< suff ]

val int : string -> (int, [ `One of int ] param_name) params_type_

val int32 : string -> (int32, [ `One of int32 ] param_name) params_type_

val int64 : string -> (int64, [ `One of int64 ] param_name) params_type_

val float : string -> (float, [ `One of float ] param_name) params_type_

val bool : string -> (bool, [ `One of bool ] param_name) params_type_

val string : string -> (string, [ `One of string ] param_name) params_type_

val file : string -> (Lib.file_info, [ `One of Lib.file_info ] param_name) params_type_

val unit : (unit, unit) params_type_

val coordinates : string -> (coordinates, [ `One of coordinates ] param_name) params_type_

val type_checker : ('a -> unit) -> ('a, 'b) params_type_ -> ('a, 'b) params_type_

val sum : ('a, 'b) params_type_ -> ('c, 'd) params_type_ -> (('a, 'c) binsum, 'b * 'd) params_type_

val prod : ('a, 'b) params_type_ -> ('c, 'd) params_type_ -> ('a * 'c, 'b * 'd) params_type_

val (**) : ('a, 'b) params_type_ -> ('c, 'd) params_type_ -> ('a * 'c, 'b * 'd) params_type_

val opt : ('a, 'b) params_type_ -> ('a option, 'b) params_type_

val neopt : ('a, 'b) params_type_ -> ('a option, 'b) params_type_

val radio : (string -> ('a, 'b) params_type_) -> string -> ('a option, 'b) params_type_

val list : string -> ('a, 'b) params_type_ -> ('a list, 'b listnames) params_type_

val set : (string -> ('a, [ `One of 'b ] param_name) params_type_) -> string -> ('a list, [ `Set of 'b ] param_name) params_type_

val any : ((string * string) list, unit) params_type_

val suffix_const : string -> (unit, [ `One of unit ] param_name) params_type_

val all_suffix : string -> (string list, [ `One of string list ] param_name) params_type_

val all_suffix_string : string -> (string, [ `One of string ] param_name) params_type_

val suffix : ?redirect_if_not_suffix:bool -> ('a, 'b) params_type_ -> ('a, 'b) params_type_

val suffix_prod : ?redirect_if_not_suffix:bool -> ('s, [< `Endsuffix | `WithoutSuffix ], 'sn) params_type -> ('a, [ `WithoutSuffix ], 'an) params_type -> ('s * 'a, [ `WithSuffix ], 'sn * 'an) params_type

val ocaml : string -> 'a Deriving_Json.t -> ('a, [ `One of 'a ocaml ] param_name) params_type_

val raw_post_data : (raw, no_param_name) params_type_

val make_list_suffix : int -> string

val make_suffix : 'a 'c. ('a, [< suff ], 'c) params_type -> 'a -> string list

val aux : 'a 'c. ('a, [< suff ], 'c) params_type -> string list option -> (string * Mod_parameters.field) list Lib.String.Table.t -> 'a -> string -> string -> (string * Mod_parameters.field) list -> string list option * (string * Mod_parameters.field) list Lib.String.Table.t * (string * Mod_parameters.field) list

val construct_params_list_raw : (string * Mod_parameters.field) list Lib.String.Table.t -> ('a, [< suff ], 'b) params_type -> 'a -> string list option * (string * Mod_parameters.field) list Lib.String.Table.t * (string * Mod_parameters.field) list

val get_to_and_of : 'a 'c. ('a, [< suff ], 'c) params_type -> 'a to_and_of

Given a parameter type, get the two functions that converts from and to strings. You should only use this function on

options ;

basic types : int, int32, int64, float, string

marshal

unit

string

bool

val walk_parameter_tree : 'a 'c. string -> ('a, [< suff ], 'c) params_type -> 'a to_and_of option

Walk the parameter tree to search for a parameter, given its name

val construct_params_string : (string * Mod_parameters.field) list -> string

val construct_params_list : (string * Mod_parameters.field) list Lib.String.Table.t -> ('a, [< suff ], 'b) params_type -> 'a -> string list option * (string * Mod_parameters.field) list

val construct_params : (string * Mod_parameters.field) list Lib.String.Table.t -> ('a, [< suff ], 'b) params_type -> 'a -> string list option * string

val make_params_names : ('a, [< suff ], 'b) params_type -> bool * 'b

val string_of_param_name : 'a -> 'a

val add_pref_params : 'a 'c. string -> ('a, [< suff ] as 'b, 'c) params_type -> ('a, 'b, 'c) params_type

val nl_prod : ('a, [< suff ] as 'b, 'an) params_type -> ('s, [ `WithoutSuffix ], 'sn) non_localized_params -> ('a * 's, 'b, 'an * 'sn) params_type

val remove_from_nlp : 'a 'c. 'b Lib.String.Table.t -> ('a, [< suff ], 'c) params_type -> 'b Lib.String.Table.t

type nl_params_set = (string * Mod_parameters.param) list Lib.String.Table.t

val empty_nl_params_set : 'a Lib.String.Table.t

val add_nl_parameter : nl_params_set -> ('a, 'b) non_localized_params_ -> 'a -> (string * Mod_parameters.field) list Lib.String.Table.t

val table_of_nl_params_set : 'a -> 'a

val list_of_nl_params_set : (string * Mod_parameters.field) list Lib.String.Table.t -> (string * Mod_parameters.field) list

val string_of_nl_params_set : (string * Mod_parameters.field) list Lib.String.Table.t -> string

val get_nl_params_names : ('a, 'b) non_localized_params_ -> 'b

val make_nlp_name : bool -> string -> string -> string

val make_non_localized_parameters : prefix:string -> name:string -> ?persistent:bool -> ('a, [ `WithoutSuffix ], 'b) params_type -> ('a, [ `WithoutSuffix ], 'b) non_localized_params

val contains_suffix : 'a 'c. ('a, [< suff ], 'c) params_type -> bool option

val wrap_param_type : 'a 'c. ('a, [< suff ] as 'b, 'c) params_type -> ('a, 'b, 'c) params_type

type _ is_unit =

| U_not : _ is_unit

| U_yes : unit is_unit

val is_unit : 'a 'c. ('a, [< suff ], 'c) params_type -> 'a is_unit

type anon_params_type = int

val anonymise_params_type : ('a, [< suff ], 'c) params_type -> anon_params_type

type files = (string * Lib.file_info) list

type +'a res_reconstr_param =

| Res_ of 'a * params' * files

| Errors_ of (string * string * exn) list * params' * files

val end_of_list : (string * 'a) list -> string -> bool

val reconstruct_params_ : ('a, [< suff ], 'b) params_type -> params' -> files -> bool -> string list option -> 'a

val reconstruct_params : sp:unit -> ('a, [< suff ], 'c) params_type -> params' Lwt.t option -> files Lwt.t option -> bool -> string list option -> 'a Lwt.t